Refactor details page
